# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.Otp:
    properties:
      email:
        type: string
      otp:
        type: integer
    required:
    - email
    type: object
...
  models.User:
    properties:
      name:
        type: string
      user_id:
        type: integer
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.UserOtp:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  models.UserSignup:
    properties:
      age:
        type: integer
      email:
        type: string
      first_name:
        type: string
      gender_id:
        type: integer
      last_name:
        type: string
      number:
        type: integer
      password:
        minLength: 6
        type: string
      user_name:
        type: string
    required:
    - email
    - password
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: SOCIAL MATCH API Gateway server.
  termsOfService: http://swagger.io/terms/
  title: Social-Match
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: User login with email and password
      parameters:
      - description: Login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - users
  /user/otp:
    post:
      consumes:
      - application/json
      description: Request OTP for user verification
      parameters:
      - description: email
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserOtp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Otp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Request OTP
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Verify user OTP
      parameters:
      - description: OTP verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Otp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify OTP
      tags:
      - users
...
  /user/signup:
    post:
      consumes:
      - application/json
      description: User signup with details
      parameters:
      - description: Signup details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSignup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Signup
      tags:
      - users
...
securityDefinitions:
  AdminAutherisation:
    in: header
    name: AccessToken
    type: apiKey
  UserAuthorization:
    in: header
    name: AccessToken
    type: apiKey
  UserConfirmToken:
    in: header
    name: ConfirmToken
    type: apiKey
swagger: "2.0"